Creation of Plan Show view

+++ db/migrate/20190211041051_create_plans.rb
@@ -2,7 +2,7 @@ class CreatePlans < ActiveRecord::Migration
   def change
     create_table :plans do |t|
       t.string  :plan_name
-      t.float   :plan_budget
+      t.integer   :plan_budget
     end
   end
 end
