+++ db/migrate/20190211041106_create_channels.rb
@@ -2,7 +2,7 @@ class CreateChannels < ActiveRecord::Migration
   def change
     create_table :channels do |t|
       t.string  :channel_name
-      t.float   :channel_budget
+      t.integer   :channel_budget
       t.integer :plan_id
     end
   end
